+++ cmd/toxiproxy.go
@@ -10,10 +10,12 @@ import (
 
 var host string
 var port string
+var config string
 
 func init() {
 	flag.StringVar(&host, "host", "localhost", "Host for toxiproxy's API to listen on")
 	flag.StringVar(&port, "port", "8474", "Port for toxiproxy's API to listen on")
+	flag.StringVar(&config, "config", "", "JSON file containing proxies to create on startup")
 	seed := flag.Int64("seed", time.Now().UTC().UnixNano(), "Seed for randomizing toxics with")
 	flag.Parse()
 	rand.Seed(*seed)
@@ -21,5 +23,8 @@ func init() {
 
 func main() {
 	server := toxiproxy.NewServer()
+	if len(config) > 0 {
+		server.PopulateConfig(config)
+	}
 	server.Listen(host, port)
 }