+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.1.0 (Unreleased)
 
+* Add -config server option to populate on startup #154
 * Updated CLI for scriptability #133
 * Add `/populate` endpoint to server #111
 * Change error responses from `title` to `error`